backend/requirements.txt
# FastAPI框架及相关
fastapi==0.100.0
uvicorn==0.22.0
pydantic==1.10.11
python-multipart==0.0.6
email-validator==2.0.0.post2

# WSGI服务器
gunicorn==21.2.0

# 数据库相关
sqlalchemy==2.0.18
alembic==1.11.1
aiosqlite==0.19.0

# 安全相关
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 缓存和会话
redis==4.5.5
aioredis==2.0.1
itsdangerous==2.1.2

# HTTP客户端
httpx==0.24.1
requests==2.31.0
aiohttp==3.8.5

# 工具库
python-dotenv==1.0.0
tenacity==8.2.2
websockets==11.0.3
orjson==3.9.1

# 时间处理
pytz==2023.3
arrow==1.2.3

# 文件处理
aiofiles==23.1.0
python-magic==0.4.27

# 日志
loguru==0.7.0

# ML相关 (算法依赖)
numpy==1.24.3
pillow==10.0.0
opencv-python==4.8.0.74

# 数据处理
pandas==2.0.3
matplotlib==3.7.2

# 系统监控
psutil==5.9.5

# 机器学习框架
# 根据系统架构安装对应版本
# 对于大多数x86/x64系统:
# torch==2.0.1
# torchvision==0.15.2
# 对于ARM架构，请使用官方指南安装: https://pytorch.org/get-started/locally/

# 监控和性能
prometheus-client==0.17.1

# 开发工具
pytest==7.4.0
black==23.3.0
isort==5.12.0
flake8==6.0.0

# 图像处理
pillow>=9.5.0

# YOLO相关
# 注意：如需使用YOLO进行目标检测，需要安装以下包
# torch>=2.0.0
# torchvision>=0.15.1
# opencv-python>=4.7.0 